Add files via upload
Hangman is a game where the user guesses the word by filling the blanks. If the user gets the word wrong in the given number of chances the hangman is executed. hangman_arts and hangman_words needs to be allocated in the same folder as hangman so to execute the program smoothly. Hangman.py is the main program. The whole code is written in Python Language